refactor: remove outdated implementation summary and related files
Deleted the outdated `IMPLEMENTATION-SUMMARY.md`, `QUERY-API.md`, and SQL migration files as they are no longer relevant. Updated the README to reflect the new fine-grained query capabilities and integrated the new API structure. The API now supports advanced query features with improved documentation and examples.
