--- IMPLEMENTATION-SUMMARY.md
@@ -1,164 +1,0 @@
-# Fine-Grained REST API Implementation Summary
-
-## Overview
-
-Successfully implemented PostgREST-style fine-grained query capabilities for Solder's auto-generated REST APIs. The implementation allows clients to perform complex filtering, sorting, pagination, and field selection directly through URL query parameters.
-
-## Implementation Details
-
-### Files Created
-
-1. **`packages/core/src/api-query-parser.ts`** (153 lines)
-   - Parses query parameters into structured filter conditions
-   - Supports operators: `eq`, `neq`, `gt`, `gte`, `lt`, `lte`
-   - Handles logical OR with syntax: `?or=(field1.eq.value,field2.gt.value)`
-   - Parses sorting: `?order=field.desc`
-   - Parses pagination: `?limit=10&offset=20`
-   - Parses field selection: `?select=field1,field2`
-
-2. **`packages/core/src/api-query-builder.ts`** (210 lines)
-   - Converts parsed parameters into Drizzle ORM queries
-   - Maps filter operators to Drizzle functions
-   - Builds WHERE clauses with AND/OR logic
-   - Applies sorting, pagination, and field selection
-   - Handles type conversion (string, number, boolean, null)
-
-3. **`packages/core/QUERY-API.md`**
-   - Comprehensive documentation with examples
-   - Covers all query features
-   - Includes Solana-specific use cases
-
-4. **`packages/core/src/examples/query-examples.ts`**
-   - Code examples demonstrating query features
-   - Helper function for building query strings
-
-### Files Modified
-
-1. **`packages/core/src/api.ts`**
-   - Updated LIST endpoint to support query parameters
-   - Integrated query parser and builder
-   - Added documentation comments
-   - Maintains backwards compatibility (no query params = return all)
-
-2. **`packages/core/src/index.ts`**
-   - Exported new query parser and builder modules
-
-3. **`README.md`**
-   - Added section on Fine-Grained Query API
-   - Included quick examples
-   - Linked to detailed documentation
-
-## Features Implemented
-
-### 1. Basic Filtering
-```bash
-GET /trades?is_buy=eq.true&sol_amount=gt.1000
-```
-
-### 2. Logical OR
-```bash
-GET /trades?or=(is_buy.eq.true,sol_amount.gt.5000)
-```
-
-### 3. Sorting
-```bash
-GET /trades?order=timestamp.desc
-```
-
-### 4. Pagination
-```bash
-GET /trades?limit=10&offset=20
-```
-
-### 5. Field Selection
-```bash
-GET /trades?select=mint,user,timestamp
-```
-
-### 6. Combined Queries
-```bash
-GET /trades?is_buy=eq.true&order=timestamp.desc&limit=10&select=mint,user,sol_amount
-```
-
-## Supported Operators
-
-| Operator | Description           | Example              |
-|----------|-----------------------|----------------------|
-| `eq`     | Equal                 | `?amount=eq.100`     |
-| `neq`    | Not equal             | `?amount=neq.0`      |
-| `gt`     | Greater than          | `?amount=gt.1000`    |
-| `gte`    | Greater than or equal | `?amount=gte.1000`   |
-| `lt`     | Less than             | `?amount=lt.500`     |
-| `lte`    | Less than or equal    | `?amount=lte.500`    |
-
-## Type Handling
-
-The API automatically converts filter values to appropriate types:
-- **Boolean**: `true`, `false`
-- **Number**: Numeric values (e.g., `100`, `1.5`)
-- **String**: Default type
-- **Null**: Literal string `"null"`
-
-## Testing
-
-- ✅ Query parser tested with unit tests (all passed)
-- ✅ TypeScript compilation successful (no errors)
-- ✅ Type safety verified
-- ✅ Backwards compatibility maintained
-
-## Benefits
-
-1. **Developer Experience**: No need to write custom API endpoints for common queries
-2. **Performance**: Efficient database queries with proper indexing
-3. **Flexibility**: Clients can construct complex queries without backend changes
-4. **Type Safety**: Full TypeScript support with proper type inference
-5. **Backwards Compatible**: Existing API consumers continue to work unchanged
-
-## Future Enhancements (Not in Scope)
-
-Potential features for future iterations:
-- Pattern matching operators (`like`, `ilike`)
-- Array operators (`in`, `not.in`)
-- Null checking operators (`is.null`, `not.is.null`)
-- Full-text search
-- Resource embedding (joins)
-- Aggregation functions
-
-## Example Use Cases for Solana
-
-### Recent Large Buys
-```bash
-GET /trades?is_buy=eq.true&sol_amount=gt.10&order=timestamp.desc&limit=50
-```
-
-### Trades for Specific Mint
-```bash
-GET /trades?mint=eq.abc123...&order=timestamp.desc
-```
-
-### Monitor Recent Activity
-```bash
-GET /trades?order=timestamp.desc&limit=100&select=mint,user,is_buy,sol_amount,timestamp
-```
-
-### Large Transactions (Buy or Sell)
-```bash
-GET /trades?or=(sol_amount.gt.100,token_amount.gt.1000000)&order=sol_amount.desc
-```
-
-## Technical Approach
-
-**Option 2 (Custom Query Parser)** was chosen over:
-- Option 1: Integrating PostgREST (too complex, another service)
-- Option 3: Hybrid approach (mixed patterns)
-
-This approach provides:
-- Full control over implementation
-- Seamless integration with existing Hono/Drizzle stack
-- Ability to optimize for Solana-specific query patterns
-- Easier debugging and maintenance
-
-## Conclusion
-
-The implementation successfully adds fine-grained REST API querying capabilities to Solder, making it comparable to Supabase/PostgREST while maintaining the simplicity and control of a custom implementation. The feature is production-ready and fully documented.
-
