--- apps/example-app/drizzle/0000_safe_mattie_franklin.sql
@@ -1,12 +1,0 @@
-CREATE TABLE "trades" (
-	"id" serial PRIMARY KEY NOT NULL,
-	"mint" varchar(44) NOT NULL,
-	"sol_amount" text NOT NULL,
-	"token_amount" text NOT NULL,
-	"is_buy" boolean NOT NULL,
-	"user" varchar(44) NOT NULL,
-	"virtual_sol_reserves" text NOT NULL,
-	"virtual_token_reserves" text NOT NULL,
-	"timestamp" timestamp NOT NULL,
-	"created_at" timestamp DEFAULT now() NOT NULL
-);
