--- apps/example-app/drizzle/meta/0000_snapshot.json
@@ -1,93 +1,0 @@
-{
-  "id": "6ef1e0a5-3549-48e5-af95-db7fbbd7a95c",
-  "prevId": "00000000-0000-0000-0000-000000000000",
-  "version": "7",
-  "dialect": "postgresql",
-  "tables": {
-    "public.trades": {
-      "name": "trades",
-      "schema": "",
-      "columns": {
-        "id": {
-          "name": "id",
-          "type": "serial",
-          "primaryKey": true,
-          "notNull": true
-        },
-        "mint": {
-          "name": "mint",
-          "type": "varchar(44)",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "sol_amount": {
-          "name": "sol_amount",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "token_amount": {
-          "name": "token_amount",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "is_buy": {
-          "name": "is_buy",
-          "type": "boolean",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "user": {
-          "name": "user",
-          "type": "varchar(44)",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "virtual_sol_reserves": {
-          "name": "virtual_sol_reserves",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "virtual_token_reserves": {
-          "name": "virtual_token_reserves",
-          "type": "text",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "timestamp": {
-          "name": "timestamp",
-          "type": "timestamp",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "created_at": {
-          "name": "created_at",
-          "type": "timestamp",
-          "primaryKey": false,
-          "notNull": true,
-          "default": "now()"
-        }
-      },
-      "indexes": {},
-      "foreignKeys": {},
-      "compositePrimaryKeys": {},
-      "uniqueConstraints": {},
-      "policies": {},
-      "checkConstraints": {},
-      "isRLSEnabled": false
-    }
-  },
-  "enums": {},
-  "schemas": {},
-  "sequences": {},
-  "roles": {},
-  "policies": {},
-  "views": {},
-  "_meta": {
-    "columns": {},
-    "schemas": {},
-    "tables": {}
-  }
-}
