--- apps/example-app/drizzle/meta/_journal.json
@@ -1,13 +1,0 @@
-{
-  "version": "7",
-  "dialect": "postgresql",
-  "entries": [
-    {
-      "idx": 0,
-      "version": "7",
-      "when": 1760332216871,
-      "tag": "0000_safe_mattie_franklin",
-      "breakpoints": true
-    }
-  ]
-}
