+++ packages/core/src/index.ts
@@ -4,6 +4,6 @@ export * from "./idl/idl-types";
 export * from "./indexer/index";
 export * from "./tables";
 export * from "./db";
-export * from "./api";
-export * from "./api-query-parser";
-export * from "./api-query-builder";
+export * from "./api/api";
+export * from "./api/api-query-parser";
+export * from "./api/api-query-builder";
